Move summary-related methods to a separate file.
