+++ scripts/backend_tests.py
@@ -30,7 +30,7 @@ import time
 
 # DEVELOPERS: Please change this number accordingly when new tests are added
 # or removed.
-EXPECTED_TEST_COUNT = 404
+EXPECTED_TEST_COUNT = 405
 
 
 COVERAGE_PATH = os.path.join(
